Refactor: Restructured Components Folder
