+++ client/components/Inputs/Buttons/Button/Button.stories.tsx
@@ -1,29 +1,23 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import PlaylistCard from './PlaylistCard';
+import Button from './Button';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
-  title: 'Cards/PlaylistCard',
-  component: PlaylistCard,
+  title: 'Buttons/Button',
+  component: Button,
   parameters: {
     layout: 'centered',
-    nextjs: {
-      appDirectory: true,
-    },
   },
+  args: { children: 'Button' },
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ['autodocs'],
   // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
   argTypes: {},
-} satisfies Meta<typeof PlaylistCard>;
+} satisfies Meta<typeof Button>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Base: Story = {
-  args: {
-    title: 'Sample Playlist',
-    owner: 'Dan',
-    imageUrl: '/images/default_playlist.png',
-  },
+export const Primary: Story = {
+  render: (args) => <Button>{args.children}</Button>,
 };
