+++ client/components/Overlay/Modals/Auth/login-modal.tsx
@@ -1,18 +1,16 @@
 'use client';
 
 import useAuthModal from '@/hooks/modals/useAuthModal';
-import { ServiceResponse } from '@/types';
-import { AXIOS } from '@/utils/network/axios';
 import * as Form from '@radix-ui/react-form';
-import { useSession } from 'next-auth/react';
+import { signIn, useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
 import React, { useEffect, useState } from 'react';
 import { toast } from 'react-hot-toast';
-import Modal from './modal';
+import Modal from '../modal';
 
-const RegisterModal = () => {
+const LoginModal = () => {
   const router = useRouter();
-  const { onClose, isOpen, isLogin, onOpen } = useAuthModal();
+  const { onClose, isOpen, isLogin } = useAuthModal();
   const [data, setData] = useState({
     username: '',
     password: '',
@@ -38,37 +36,24 @@ const RegisterModal = () => {
 
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    try {
-      const { data: res } = await AXIOS.post<ServiceResponse<number>>(
-        '/auth/register',
-        {
-          username: data.username,
-          password: data.password,
-        }
-      );
-      if (res.success) {
-        toast.success(`Registered account, please login!`);
-        setTimeout(() => onOpen(true), 350);
-      }
-      /* eslint-disable @typescript-eslint/no-explicit-any */
-    } catch (err: any) {
-      const message = err?.response?.data?.message ?? 'Unexpected Error';
-
-      if (err.response.status === 409) {
-        toast(message, {
-          icon: '⚠️',
-        });
-      } else {
-        toast.error(message);
-      }
+    const res = await signIn('credentials', {
+      ...data,
+      redirect: false,
+      // callbackUrl: '/',
+    });
+    if (res?.error === 'CredentialsSignin') {
+      toast.error('Invalid Credentials Provided');
+    } else {
+      toast.success(`Welcome back`);
+      // onClose();
     }
   };
 
   return (
     <Modal
-      title="Create a new account"
-      description="Enter your details"
-      isOpen={isOpen && !isLogin}
+      title="Welcome back"
+      description="Login to your account"
+      isOpen={isOpen && isLogin}
       onChange={onChange}
     >
       <Form.Root onSubmit={handleSubmit} className="">
@@ -86,7 +71,7 @@ const RegisterModal = () => {
           </div>
           <Form.Control asChild>
             <input
-              placeholder="Enter a username"
+              placeholder="Your Username"
               className="shadow-blackA9 box-border inline-flex h-[35px] w-full appearance-none items-center justify-center rounded-[4px] bg-neutral-700 px-[10px] text-[15px] leading-none text-white outline-none placeholder:text-sm placeholder:font-extralight placeholder:text-neutral-500"
               type="username"
               value={data.username}
@@ -109,7 +94,7 @@ const RegisterModal = () => {
           </div>
           <Form.Control asChild>
             <input
-              placeholder="Enter a secure password"
+              placeholder="Your Password"
               value={data.password}
               onChange={(e) => setData({ ...data, password: e.target.value })}
               className="shadow-blackA9 box-border inline-flex h-[35px] w-full appearance-none items-center justify-center rounded-[4px] bg-neutral-700 px-[10px] text-[15px] leading-none text-white outline-none placeholder:text-sm placeholder:font-extralight placeholder:text-neutral-500"
@@ -119,8 +104,8 @@ const RegisterModal = () => {
           </Form.Control>
         </Form.Field>
         <Form.Submit asChild>
-          <button className="shadow-blackA7 mt-[10px] box-border inline-flex h-[35px] w-full items-center justify-center rounded-[4px] bg-white px-[15px] font-medium leading-none text-black shadow-[0_2px_10px] hover:bg-mauve3 focus:shadow-[0_0_0_2px] focus:shadow-black focus:outline-none">
-            Register
+          <button className="shadow-blackA7 mt-[10px] box-border inline-flex h-[35px] w-full items-center justify-center rounded-[4px] bg-white px-[15px] font-medium leading-none text-black text-violet11 shadow-[0_2px_10px] hover:bg-mauve3 focus:shadow-[0_0_0_2px] focus:shadow-black focus:outline-none">
+            Login
           </button>
         </Form.Submit>
       </Form.Root>
@@ -128,4 +113,4 @@ const RegisterModal = () => {
   );
 };
 
-export default RegisterModal;
+export default LoginModal;
