+++ client/components/Overlay/Modals/MultiUploadModal/upload-song.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { revalidatePathClient } from '@/app/action';
-import Input from '@/components/input';
+import Input from '@/components/Inputs/Input/input';
 import { getDefaultImage } from '@/utils';
 import { AXIOS } from '@/utils/network/axios';
 import * as Dialog from '@radix-ui/react-dialog';
