+++ client/components/Utils/RedirectSync.tsx/RedirectSync.tsx
